HIVE-26455: Remove PowerMock from hive-exec
